Changes by Gebruiker

+++ src/main/java/crane/Crane.java
@@ -29,7 +29,7 @@ public class Crane {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
@@ -37,11 +37,11 @@ public class Crane {
         return holdingContainer;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
